HTML and Vue3 landing page for a musical festival.
This project was a funny break to me back when I worked for ZJ Events LLC. It was originally required to be created on top of a Wordpress installation. I decided to go the HTML - CSS- Bootstrap - Vue3 way, instead, accomplishing it in a fraction of the estimated time and also meeting 100% the graphic design goals and maximum performance. Actual sponsors and contact urls have been stripped from these files, in order to avoid property right issues.